feat: implement DummyBackend for metadata extraction and update routes
