+++ bibra/backend/__init__.py
@@ -1,0 +1,3 @@
+from bibra.backend.dummy import DummyBackend
+
+__all__ = ["DummyBackend"]
